examples: linux: Factor out lookup_channel into common file
