+++ examples/linux/rpmsg-proxy-app/proxy_app.c
@@ -1,6 +1,5 @@
 //SPDX-License-Identifier: BSD-3-Clause
 
-#include <dirent.h>
 #include <errno.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -206,50 +205,6 @@ int file_write(char *path, char *str)
 void exit_action_handler(int signum)
 {
 	proxy->active = 0;
-}
-
-static void set_src_dst(char *out, struct rpmsg_endpoint_info *pep)
-{
-	long dst = 0;
-	char *lastdot = strrchr(out, '.');
-
-	if (lastdot == NULL)
-		return;
-	dst = strtol(lastdot + 1, NULL, 10);
-	if ((errno == ERANGE && (dst == LONG_MAX || dst == LONG_MIN))
-	    || (errno != 0 && dst == 0)) {
-		return;
-	}
-	pep->dst = (unsigned int)dst;
-}
-
-/*
- * return the first dirent matching rpmsg-openamp-demo-channel
- * in /sys/bus/rpmsg/devices/ E.g.:
- *	virtio0.rpmsg-openamp-demo-channel.-1.1024
- */
-static int lookup_channel(char *out, struct rpmsg_endpoint_info *pep)
-{
-	char dpath[] = RPMSG_BUS_SYS "/devices";
-	struct dirent *ent;
-	DIR *dir = opendir(dpath);
-
-	if (dir == NULL) {
-		fprintf(stderr, "opendir %s, %s\n", dpath, strerror(errno));
-		return -EINVAL;
-	}
-	while ((ent = readdir(dir)) != NULL) {
-		if (strstr(ent->d_name, pep->name)) {
-			strncpy(out, ent->d_name, NAME_MAX);
-			set_src_dst(out, pep);
-			printf("using dev file: %s\n", out);
-			closedir(dir);
-			return 0;
-		}
-	}
-	closedir(dir);
-	fprintf(stderr, "No dev file for %s in %s\n", pep->name, dpath);
-	return -EINVAL;
 }
 
 void kill_action_handler(int signum)
